fix: styling and removed Hello World from App.tsx
