+++ frontend/src/components/SpotifySearch.tsx
@@ -1,75 +1,12 @@
+import { useState } from 'react'
 import { useQuery } from '@tanstack/react-query'
 import { MagnifyingGlassIcon } from '@heroicons/react/24/outline'
-import React, { useEffect, useState } from 'react'
 
-import { Track } from 'lib/types'
 import { searchTracks } from 'lib/api'
-
-// import { getRecommendedTracks } from 'lib/api'
-// import Loading from 'components/Loading'
-
-// type Inputs = {
-//   example: string
-//   exampleRequired: string
-// }
-
-// const SpotifySearch: React.FC = () => {
-//   const {
-//     register,
-//     handleSubmit,
-//     watch
-//     // formState: { errors }
-//   } = useForm<Inputs>()
-//   const onSubmit: SubmitHandler<Inputs> = (data) => {
-//     console.log('data', data)
-//   }
-
-//   console.log(watch('example'))
-
-//   return (
-//     <div>
-//       <form className="my-3" onSubmit={handleSubmit(onSubmit)}>
-//         <div className="flex items-center rounded-full border border-neutral-600 px-3 py-1">
-//           <MagnifyingGlassIcon className="h-5 w-5 text-neutral-600" />
-//           <input
-//             className="bg-inherit px-2 py-1 placeholder:text-neutral-300 focus:outline-none dark:placeholder:text-neutral-500"
-//             placeholder="Search"
-//             autoComplete="off"
-//             {...register('example')}
-//           />
-//         </div>
-
-//         {/* <input
-//         className="rounded-md border px-2 py-1"
-//         {...register('exampleRequired', { required: true })}
-//       /> */}
-//         {/* errors will return when field validation fails  */}
-//         {/* {errors.exampleRequired && <span>This field is required</span>} */}
-
-//         {/* <button type="submit">Search</button> */}
-//       </form>
-//       {/* <button>Search</button> */}
-//       {/* <Loading /> */}
-//     </div>
-//   )
-// }
-
-import { useCallback } from 'react'
-import TrackPreview from 'components/TrackPreview'
+import { Track } from 'lib/types'
+import { useDebounce } from 'lib/hooks'
 import Loading from 'components/Loading'
-
-function useDebounceEffect(
-  effect: any,
-  deps: React.DependencyList,
-  delay = 250
-) {
-  const callback = useCallback(effect, [...deps, effect])
-
-  useEffect(() => {
-    const timeout = setTimeout(callback, delay)
-    return () => clearTimeout(timeout)
-  }, [callback, delay])
-}
+import TrackPreview from 'components/TrackPreview'
 
 interface InputProps {
   setValue: React.Dispatch<string>
@@ -77,7 +14,7 @@ interface InputProps {
 
 const Input: React.FC<InputProps> = ({ setValue }) => {
   const [search, setSearch] = useState('')
-  useDebounceEffect(
+  useDebounce(
     () => {
       setValue(search)
     },
@@ -87,16 +24,16 @@ const Input: React.FC<InputProps> = ({ setValue }) => {
 
   return (
     <>
-      <div className="group my-2 flex items-center rounded-full border border-neutral-600 px-3 py-1 focus-within:border-neutral-500 focus-within:bg-neutral-800">
+      <div className="group my-2 flex items-center rounded-full border px-3 py-1 dark:border-neutral-500 dark:focus-within:border-neutral-500 dark:focus-within:bg-neutral-800">
         <MagnifyingGlassIcon
-          className={`h-5 w-5 text-neutral-600 ${
+          className={`h-5 w-5 ${
             search.length > 0
               ? 'text-neutral-50'
-              : 'group-focus-within:text-neutral-500'
+              : 'text-neutral-300 group-focus-within:text-neutral-500 dark:text-neutral-600'
           }`}
         />
         <input
-          className="select-none bg-inherit px-2 py-1 placeholder:text-neutral-300 focus:outline-none dark:placeholder:text-neutral-500"
+          className="w-full select-none bg-inherit px-2 py-1 placeholder:text-neutral-300 focus:outline-none group-focus-within:placeholder:text-neutral-500 dark:placeholder:text-neutral-600"
           placeholder="Search tracks on Spotify"
           autoComplete="off"
           value={search}
@@ -138,17 +75,17 @@ const SpotifySearch: React.FC<Props> = ({ setChosen }) => {
         {search && (
           <div className="rounded-lg py-1.5 dark:bg-neutral-800">
             {isLoading ? (
-              <div className="px-3.5">
+              <div className="px-3">
                 <Loading className="my-1" />
               </div>
             ) : isError ? (
-              <div>Error</div>
+              <div className="px-3">Error</div>
             ) : data ? (
               <div>
                 <div>
                   {data.map((track: Track) => (
                     <TrackPreview
-                      className="cursor-pointer select-none px-3.5 hover:bg-neutral-700"
+                      className="cursor-pointer select-none px-3 hover:bg-neutral-700"
                       key={track.uri}
                       track={track}
                       onClick={() => {
