+++ frontend/src/components/TrackPreview.tsx
@@ -21,8 +21,13 @@ const TrackPreview: React.FC<Props> = ({ className, track, onClick }) => {
         />
       )}
       <div className="min-w-0">
-        <p className="overflow-hidden overflow-ellipsis whitespace-nowrap">
-          {track.name}
+        <p className="flex items-center overflow-hidden overflow-ellipsis whitespace-nowrap">
+          {track.name}{' '}
+          {!!track.explicit && (
+            <span className="ml-1.5 flex items-center rounded-sm px-1 py-[3px] text-[10px] font-medium leading-none dark:bg-neutral-300 dark:text-neutral-900">
+              E
+            </span>
+          )}
         </p>
         <div className="overflow-hidden overflow-ellipsis text-sm dark:text-neutral-400/90">
           {track.artists.map((artist: any) => (
